SPM-827 Isolate "segmentio/kafka-go" usage to mqueue only
